+++ client/src/components/pages/HomePage.tsx
@@ -1,14 +1,11 @@
-import React, { useEffect, useState } from "react";
-import { useAuth0 } from "@auth0/auth0-react";
-import { LogoutButton } from "../Auth0/LogoutButton";
-import { useNavigate } from "react-router-dom";
-import axios from "axios";
-import ButtonToLobby from '../buttons/ButtonToLobby';
-import ButtonToFindGame from '../buttons/ButtonToFindGame'
-import ButtonToProfile from '../buttons/ButtonToProfile'
+import React, { useEffect, useState } from 'react';
+import { useAuth0 } from '@auth0/auth0-react';
+import { LogoutButton } from '../Auth0/LogoutButton';
+import { useNavigate } from 'react-router-dom';
+import axios from 'axios';
+import { ButtonToProfile, ButtonToFindGame, ButtonToLobby } from '../Buttons';
 
 // import ChaseCam from '../components/ChaseCam'
-
 
 interface UserData {
   username: string;
@@ -22,12 +19,11 @@ const HomePage = () => {
   const [userData, setUserData] = useState<UserData | null>(null);
   const navigate = useNavigate();
 
-
   useEffect(() => {
     const fetchUserData = async () => {
       try {
         // Check if the user exists by sending a POST request instead of a GET request
-        const response = await axios.post<UserData>("/Users", {
+        const response = await axios.post<UserData>('/Users', {
           username: user?.name,
           email: user?.email,
           authId: user?.sub,
@@ -36,7 +32,7 @@ const HomePage = () => {
         setUserData(response.data);
         // console.log(response);
       } catch (error) {
-        console.error("Error fetching user data:", error);
+        console.error('Error fetching user data:', error);
       }
     };
 
@@ -55,9 +51,8 @@ const HomePage = () => {
         <h1>{`Welcome Home, ${user.given_name}`}!</h1>
         <ButtonToProfile />
         <ButtonToFindGame />
-        <LogoutButton />
         <ButtonToLobby />
-        <LogoutButton/>
+        <LogoutButton />
         {/* <ChaseCam /> */}
       </div>
     )
